@@ -27,6 +27,7 @@
     "what-input": "5.0.5"
   },
   "devDependencies": {
+    "@types/node": "^18.19.115",
     "@types/react": "16.1.0",
     "@types/react-dom": "16.0.4",
     "@types/react-hot-loader": "3.0.6",
@@ -37,30 +38,40 @@
     "@types/react-transition-group": "2.0.8",
     "@types/valid-url": "1.0.2",
     "@types/webpack-env": "1.13.5",
+    "@typescript-eslint/eslint-plugin": "^8.36.0",
+    "@typescript-eslint/parser": "^8.36.0",
     "babel-loader": "7.1.4",
     "chalk": "2.3.2",
     "clean-webpack-plugin": "0.1.19",
     "copy-webpack-plugin": "4.5.1",
+    "cross-env": "^7.0.3",
     "css-loader": "0.28.11",
+    "eslint": "^9.30.1",
+    "eslint-config-prettier": "^10.1.5",
+    "eslint-plugin-import": "^2.32.0",
+    "eslint-plugin-jsx-a11y": "^6.10.2",
+    "eslint-plugin-react": "^7.37.5",
+    "eslint-plugin-react-hooks": "^5.2.0",
     "express": "4.16.3",
     "extract-text-webpack-plugin": "4.0.0-beta.0",
     "favicons-webpack-plugin": "0.0.9",
     "file-loader": "1.1.11",
+    "globals": "^16.3.0",
     "html-webpack-plugin": "3.0.6",
-    "husky": "0.14.3",
+    "husky": "^9.1.7",
     "lint-staged": "7.0.4",
-    "sass": "^1.73.0",
+    "prettier": "^3.6.2",
     "redux-devtools-extension": "2.13.2",
     "resolve-url-loader": "2.3.0",
+    "sass": "^1.73.0",
     "sass-loader": "6.0.7",
     "style-loader": "0.20.3",
+    "terser-webpack-plugin": "^2.3.8",
     "ts-loader": "4.0.1",
-    "tslint": "5.9.1",
     "tslint-config-prettier": "1.10.0",
-    "tslint-microsoft-contrib": "5.0.3",
     "tslint-react": "3.5.1",
-    "typescript": "2.7.2",
-    "uglifyjs-webpack-plugin": "1.2.4",
+    "typescript": "^5.8.3",
+    "typescript-eslint": "^8.36.0",
     "webpack": "4.1.1",
     "webpack-cli": "2.0.12",
     "webpack-dev-server": "3.1.1",
@@ -68,25 +79,25 @@
     "webpack-merge": "4.1.2",
     "webpackbar": "2.6.1"
   },
+  "overrides": {
+    "node-sass": "npm:sass@^1.73.0"
+  },
   "scripts": {
-    "tscheck": "tsc --noEmit",
-    "tslint": "tslint --project .",
-    "prepush": "npm run tscheck",
-    "precommit": "lint-staged",
+    "lint": "eslint . --ext .ts,.tsx",
+    "lint:fix": "eslint . --ext .ts,.tsx --fix",
+    "prepare": "husky install",
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "webpack-dev-server --mode development --config ./webpack_config/webpack.dev.js",
-    "build": "webpack --mode production --config webpack_config/webpack.prod.js -p",
+    "build": "NODE_OPTIONS=--openssl-legacy-provider webpack --mode production --config webpack_config/webpack.prod.js -p",
     "prettier:diff": "prettier --write --config ./.prettierrc --list-different \"src/**/*.ts\" \"src/**/*.tsx\"",
     "formatAll": "find ./src/ -name '*.ts*' | xargs prettier --write --config ./.prettierrc --config-precedence file-override"
   },
-  "precommit": "lint-staged",
   "lint-staged": {
     "*.{ts,tsx}": [
       "prettier --write --config ./.prettierrc --config-precedence file-override",
       "git add"
     ]
   },
-  "prepush": "npm run tslint && npm run tscheck",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/MyCryptoHQ/React-Typescript-Web3.git"
